requirements.txt
phidata
python-dotenv
packaging
fastapi
uvicorn
groq
psycopg[binary]
pypdf
sqlalchemy
google-generativeai